requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=10.1.0
opencv-python>=4.8.1
numpy>=1.26.2
zxing-cpp>=2.3.0
